+++ src/commands/settings.ts
@@ -1,0 +1,86 @@
+import discord, { StringSelectMenuBuilder } from 'discord.js'
+import type ApplicationCommand from './ApplicationCommand'
+import userController from '../controllers/users'
+
+async function createInteractionReply(userId: string) {
+  const user = await userController.get(userId)
+
+  return {
+    content: 'Configure the bot with the options below.',
+    ephemeral: true,
+    components: [
+      new discord.ActionRowBuilder<discord.ButtonBuilder>().addComponents(
+        new discord.ButtonBuilder()
+          .setCustomId('verseDisplay')
+          .setLabel(
+            user.preferences.verseDisplay === 'embed'
+              ? 'Display Mode: Embed'
+              : 'Display Mode: Blockquote'
+          )
+          .setStyle(
+            user.preferences.verseDisplay === 'embed'
+              ? discord.ButtonStyle.Primary
+              : discord.ButtonStyle.Secondary
+          ),
+        new discord.ButtonBuilder()
+          .setCustomId('inlineVerses')
+          .setLabel(
+            `Inline Verses: ${user.preferences.inlineVerses ? 'On' : 'Off'}`
+          )
+          .setStyle(
+            user.preferences.inlineVerses
+              ? discord.ButtonStyle.Success
+              : discord.ButtonStyle.Danger
+          )
+      )
+    ]
+  }
+}
+
+async function awaitMessageComponent(
+  userId: string,
+  response: discord.InteractionResponse
+) {
+  const user = await userController.get(userId)
+
+  await response
+
+  response
+    .awaitMessageComponent({
+      filter: i => i.user.id === userId
+    })
+    .then(async i => {
+      if (i.customId === 'verseDisplay') {
+        if (user.preferences.verseDisplay === 'embed') {
+          user.preferences.verseDisplay = 'blockquote'
+        } else {
+          user.preferences.verseDisplay = 'embed'
+        }
+      } else if (i.customId === 'inlineVerses') {
+        user.preferences.inlineVerses = !user.preferences.inlineVerses
+      }
+
+      await user.save()
+
+      awaitMessageComponent(
+        user.id,
+        await i.update(await createInteractionReply(user.id))
+      )
+    })
+}
+
+const command: ApplicationCommand = {
+  meta: {
+    name: 'settings',
+    description: 'Configure the bot to your liking.'
+  },
+  execute: async interaction => {
+    const response = await interaction.reply(
+      await createInteractionReply(interaction.user.id)
+    )
+
+    awaitMessageComponent(interaction.user.id, response)
+  }
+}
+
+export default command