+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `/changelog` command to show this file.
+- `/settings` to configure the bot.
 - set nickname to `PBDB` when joining a server.
 - post Bible verses referenced in messages.